Introduced Support for TLS
Added a new boolean property called IsTlsRequired. If set to true, it uses an SslStream to write the syslog data to the end point rather than the TcpClient directly.

Had to change the RemoteAddress property to be a string to support both IP addresses and Host Names. Even when specifying a host name previously it would get resolved to an IP address and the server certificate validation would fail. Not sure if this would break anything else that I haven't tested, like broadcast addresses.

If you decide to incorporate this, let me know so I can replace my custom implementation with the official version.

Hope this is useful.